account_service/app/api/core/security.py
from jose import jwt, JWTError
from fastapi import HTTPException, status
from datetime import datetime, timedelta
from app.api.core.config import settings


def decode_access_token(token: str) -> dict:
    try:
        payload = jwt.decode(token, settings.jwt_secret, algorithms=[settings.jwt_alg])
        return payload
    except JWTError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid authentication credentials",
            headers={"WWW-Authenticate": "Bearer"},
        )
